# openapi.yaml
openapi: 3.0.3
info:
  title: Control of Moedinhas API
  description: API para gerenciamento de domicílios, usuários, contas, categorias, transações e divisão de despesas.
  version: "1.0.0"
servers:
  - url: http://localhost:8000
paths:
  /users/:
    post:
      summary: Criar usuário
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreatedResponse'
    get:
      summary: Listar usuários
      tags: [Users]
      parameters:
        - name: household_id
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/{user_id}:
    get:
      summary: Obter usuário
      tags: [Users]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
...
    UserCreate:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        household_id:
          type: string
    UserCreatedResponse:
      type: object
      properties:
        message:
          type: string
        user_id:
          type: string
        name:
          type: string
        email:
          type: string
        household_id:
          type: string
...
security: []